Extract an idea form into a single component
Closes #18
